requirements.txt
aiofiles==24.1.0
aiohttp==3.12.15
aioshutil==1.5
albumentations==2.0.8
arabic_reshaper==3.0.0
backports.statistics==0.1.0
colorama==0.4.6
ctranslate2==3.24.0
customtkinter==5.2.2
einops==0.8.1
fastapi==0.116.1
freetype_py==2.5.1
googletrans==4.0.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.34.4
ijson==3.4.0
kornia==0.8.1
langcodes==3.5.0
loguru==0.7.3
manga_ocr==0.1.14
networkx==3.3
numpy==2.3.2
omegaconf==2.3.0
onnxruntime==1.22.1
openai==1.102.0
opencv_python==4.12.0.88
packaging==25.0
pandas==2.3.2
Pillow==11.3.0
psutil==5.9.5
py3langid==0.2.2
pyclipper==1.3.0.post6
pydantic==2.11.7
pydensecrf==1.0
PyHyphen==4.0.4
PySide6==6.9.2
pytest==7.2.2
python-dotenv==1.1.1
pytorch_lightning==2.5.4
regex==2025.7.34
Requests==2.32.5
rich==14.1.0
rusty_manga_image_translator==0.8.15
safetensors==0.6.2
scipy==1.16.1
sentencepiece==0.2.1
setuptools==70.2.0
Shapely==2.1.1
skimage==0.0
starlette==0.47.3
tiktoken==0.11.0
timm==1.0.19
torch==2.8.0+cu129
torch_summary==1.4.5
torchvision==0.23.0+cu129
tqdm==4.67.1
uvicorn==0.35.0
websockets==14.2
xformers==0.0.32.post2
